example/pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Sheena Artrip <sheena.artrip@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.0"
funcy = "^1.17"
django-dramatiq = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
